+++ Sources/ThemeKitSwiftUI/Extensions/SystemColorScheme+ColorScheme.swift
@@ -30,7 +30,7 @@ extension ColorScheme {
   }
 }
 
-#if canImport(UIKit)
+#if os(iOS)
 extension UIUserInterfaceStyle {
   /// Creates a `UIUserInterfaceStyle` from a SwiftUI `ColorScheme`.
   public init(_ colorScheme: ColorScheme) {